requirements.txt
Flask
flask-cors
sentence-transformers
psycopg2-binary
pgvector
